build: check if on jlab system to load central external source. Better and safer set_enviornment script

+++ set_environment.sh
@@ -1,54 +1,96 @@
-# launched automatically by conda activate.d
-echo "Sourcing additional enviornment variables"
+echo ""
+echo "*******************"
+echo "Loading environment"
+echo "*******************"
+echo ""
+
+env_name="pyamptools"
 
 ####################
-### IF ON JLAB IFARM ###
-module add cuda
-export CUDA_INSTALL_PATH=/apps/cuda/11.4.2/
-export GPU_ARCH=sm_75
+# Check if the hostname contains "jlab.org" if so we perform default setup
+#    and link external libraries against a central repo (its actually my
+#    personal working directory but it isn't modified much anymore)
+
+hostname=$(hostname)
+if [[ "$hostname" == *"jlab.org"* ]]; then
+
+    echo "Hostname contains 'jlab.org'. Loading JLab environment..."
+    echo ""
+
+    # Was in .bashrc is this good place now?
+    source /etc/profile.d/modules.sh
+    module add cuda
+    module use /apps/modulefiles
+    module load mpi/openmpi3-x86_64
+    module load gcc/9.3.0
+
+    export CUDA_INSTALL_PATH=/apps/cuda/11.4.2/
+    export GPU_ARCH=sm_75
+    export LD_LIBRARY_PATH=/usr/lib64:$LD_LIBRARY_PATH # needed for mpi4py
+
+    # You could build the external libraries from source but since you are
+    # on the JLab system you can also just use the pre-built libraries
+    mv external .external # hide default source
+    ln -s /w/halld-scshelf2101/lng/WORK/PyAmpTools/external .
+fi
 ####################
 
-# Check if REPO_HOME is a full path OR if REPO_HOME is unset or empty
+# Perform some checks if you are currently in the base conda environment
+if [ -z "$CONDA_DEFAULT_ENV" ] || [ "$CONDA_DEFAULT_ENV" == "base" ]; then
+    echo "You are currently in the base conda environment (or conda doesn't exist at all)"
+    if conda env list | grep -q "$env_name"; then
+        echo "Please activate '$env_name' environment and SOURCE AGAIN"
+    else
+        # of course I can create and activate the env here but the user should manually do it (I think)
+        echo "Conda environment '$env_name' does not exist. Please create it, activate it, and SOURCE AGAIN"
+    fi
+    echo ""
+    return
+fi
 # VSCode could create additional environment variables... Checking to see if
-#   REPO_HOME is a full path is another check to see if it's been set by the user
-if [[ ! $REPO_HOME == /* ]] || [ -z "$REPO_HOME" ]; then
-    export REPO_HOME=$( dirname "$( readlink -f "${BASH_SOURCE[0]}" )" ) # absolute location of this script
-    echo "Setting REPO_HOME to $REPO_HOME"
+#   PYAMPTOOLS_HOME is a full path is another check to see if it's been set by the user
+if [[ ! $PYAMPTOOLS_HOME == /* ]] || [ -z "$PYAMPTOOLS_HOME" ]; then
+    export PYAMPTOOLS_HOME=$( dirname "$( readlink -f "${BASH_SOURCE[0]}" )" ) # absolute location of this script
+    echo "Setting PYAMPTOOLS_HOME to $PYAMPTOOLS_HOME"
 else
-    echo "REPO_HOME is already set. Will not attempt override. REPO_HOME=$REPO_HOME"
+    echo "PYAMPTOOLS_HOME is already set. Will not attempt override. PYAMPTOOLS_HOME=$PYAMPTOOLS_HOME"
 fi
 
-export AMPTOOLS_HOME=$REPO_HOME/external/AmpTools
+export AMPTOOLS_HOME=$PYAMPTOOLS_HOME/external/AmpTools
 export AMPTOOLS=$AMPTOOLS_HOME/AmpTools
 export AMPPLOTTER=$AMPTOOLS_HOME/AmpPlotter
-export FSROOT=$REPO_HOME/external/FSRoot
+export FSROOT=$PYAMPTOOLS_HOME/external/FSRoot
 
 ## Load Generator library if external/AMPTOOLS_GENERATORS was built
-if [ -d "$REPO_HOME/external/AMPTOOLS_GENERATORS/ccdb" ]; then
-    # export CCDB_HOME=$REPO_HOME/external/AMPTOOLS_GENERATORS/ccdb
+if [ -d "$PYAMPTOOLS_HOME/external/AMPTOOLS_GENERATORS/ccdb" ]; then
+    # export CCDB_HOME=$PYAMPTOOLS_HOME/external/AMPTOOLS_GENERATORS/ccdb
     # source $CCDB_HOME/environment.bash
-    export LD_LIBRARY_PATH=$REPO_HOME/external/AMPTOOLS_GENERATORS:$LD_LIBRARY_PATH
+    export LD_LIBRARY_PATH=$PYAMPTOOLS_HOME/external/AMPTOOLS_GENERATORS:$LD_LIBRARY_PATH
 fi
 
 export LD_LIBRARY_PATH=$LD_LIBRARY_PATH  # for libhwloc.so.5 needed to locate hardware for openmpi
 export LD_LIBRARY_PATH=$AMPTOOLS/lib:$LD_LIBRARY_PATH
 export LD_LIBRARY_PATH=$AMPTOOLS_HOME/AmpPlotter/lib:$LD_LIBRARY_PATH
 export LD_LIBRARY_PATH=$FSROOT/lib:$LD_LIBRARY_PATH
-export LD_LIBRARY_PATH=$REPO_HOME/external/AMPTOOLS_AMPS_DATAIO:$LD_LIBRARY_PATH
+export LD_LIBRARY_PATH=$PYAMPTOOLS_HOME/external/AMPTOOLS_AMPS_DATAIO:$LD_LIBRARY_PATH
 
 ##################### Activate ROOT #################
-if [ -f "$REPO_HOME/external/root/thisroot.sh" ]; then
-    source $REPO_HOME/external/root/thisroot.sh $REPO_HOME/external/root # setup ROOT
+if [ -f "$PYAMPTOOLS_HOME/external/root/thisroot.sh" ]; then
+    source $PYAMPTOOLS_HOME/external/root/thisroot.sh $PYAMPTOOLS_HOME/external/root # setup ROOT
 else
     echo ""
     echo "ROOT not found. Please go into root and run build_root.sh to your specifications."
 fi
 
-python $REPO_HOME/utility/link_modules.py # symlink C header files so it can be found by ROOT
+python $PYAMPTOOLS_HOME/utility/link_modules.py # symlink C header files so it can be found by ROOT
 
 # Register pa (pyamptools dispatch system) for tab auto-completion of avaiable commands
 eval "$(register-python-argcomplete pa)"
 
 # setup auto-load for next time
 mkdir -p $CONDA_PREFIX/etc/conda/activate.d/
-ln -sf "$(readlink -f set_environment.sh)" $CONDA_PREFIX/etc/conda/activate.d/
+# readlink -f can still return a location even if set_environment.sh is not in the current directory
+#    lets check that it is a valid file first
+if [ -e "$(readlink -f set_environment.sh)" ]; then
+    ln -sf "$(readlink -f set_environment.sh)" $CONDA_PREFIX/etc/conda/activate.d/
+fi
